♻️ UserType 정의 및 User API 일부 로직 수정
