+++ src/query-hooks/user/key.ts
@@ -1,7 +1,7 @@
-export const QUERY_KEY = {
+export const USER_QUERY_KEY = {
     base: ['user'],
     info: (userId?: string) =>
         userId
-            ? [...QUERY_KEY.base, 'info', userId]
-            : [...QUERY_KEY.base, 'info'],
+            ? [...USER_QUERY_KEY.base, 'info', userId]
+            : [...USER_QUERY_KEY.base, 'info'],
 };
