+++ src/types/auth.ts
@@ -1,0 +1,8 @@
+export type SocialPlatformType = 'apple' | 'kakao';
+
+export interface UserType {
+    userId: string;
+    email: string;
+    username: string;
+    imageUrl: string;
+}
