+++ src/types/index.ts
@@ -1,4 +1,4 @@
-export * from './login';
+export * from './auth';
 export * from './map';
 export * from './course';
 export * from './search';
