--- src/types/login.ts
@@ -1,1 +1,0 @@
-export type SocialPlatformType = 'apple' | 'kakao';
